Fixed an issue with UpdateSidechain for non-ceasable sidechains
UpdateSidechain() can be called several times for non-ceasable sidechains if in a block there is more than one certificate (as oppose to ceasable ones that call the function only once).

If a block included for a non-ceasable sidechain two certificates, the first of epoch 0 and the other of epoch 1, the undo data were wrongly overwritten while processing the second certificate.