Extended BlockchainHelper to handle referenced_height for certificates
Now the Python BlockchainHelper automatically handles the `referencedHeight` parameter when generating certificates for non-ceasing sidechains.

Basically, for every certificate to be generated, it tries to use the lowest height possible (starting from the block after the creation of the sidechain itself) to increase the probability to have colissions in a block (where a collision is the situation in which we want several certificates in a block but they don't have enough blocks to make a reference).

It is also possible to manually specify a referenced height when creating the certificate.