Fixed the getcertmaturityinfo RPC command for non-ceasable sidechains
In case we had several certificates in the mempool for a non-ceasable sidechain, the `getcertmaturityinfo` RPC command was returning "TOP_QUALITY" only for the "top epoch" one, and "LOW_QUALITY" for the other ones.

Now it's returning "TOP_QUALITY" for every certificate related to non-ceasable sidechains.

sc_getcertmaturityinfo.py has been finalized to include checks for this new behavior.