Enforce prev cert inclusion for v2 sidechain certificates
This commit changes the rules for non ceasing sidechain certificates, so that each certificate must reference a block
whose height is greater or equal to the height of the block containing the previous certificate.
This is the fundamental fix for the "ban problem". This also implies that it is never possible to send more than one
certificate per block (hence the certificate rate is limited).