Removed test that checks fees for multiple certs in a block
We updated the Python test that checked how sidechain fees are updated.
In particular, the final section of the test for non-ceasible v2 sc has been removed
as it implemented a situation where more than a certificate is sent on a single
block, which is now no longer valdi with the recent update.